+++ backend/src/main/java/com/example/backend/member/dto/MemberForm.java
@@ -43,7 +43,7 @@ public class MemberForm {
 
     @NotBlank(message = "주소는 필수입니다.")
     private String address;
-    
+
     @NotNull(message = "생년월일은 필수입니다.")
     @Past(message = "생년월일은 과거 날짜여야 합니다.")
     private LocalDate birthday;
