+++ backend/src/main/java/com/example/backend/product/dto/ProductDto.java
@@ -1,4 +1,4 @@
-package com.example.backend.dto;
+package com.example.backend.product.dto;
 
 import lombok.Data;
