+++ backend/src/main/java/com/example/backend/product/entity/Product.java
@@ -1,4 +1,4 @@
-package com.example.backend.entity;
+package com.example.backend.product.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
