+++ backend/src/main/java/com/example/backend/product/repository/ProductImageRepository.java
@@ -1,6 +1,6 @@
-package com.example.backend.repository;
+package com.example.backend.product.repository;
 
-import com.example.backend.entity.ProductImage;
+import com.example.backend.product.entity.ProductImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductImageRepository extends JpaRepository<ProductImage, Integer> {
