+++ backend/src/main/java/com/example/backend/product/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package com.example.backend.repository;
+package com.example.backend.product.repository;
 
-import com.example.backend.entity.Product;
+import com.example.backend.product.entity.Product;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
