+++ backend/src/main/java/com/example/backend/product/service/ProductService.java
@@ -1,12 +1,12 @@
-package com.example.backend.service;
+package com.example.backend.product.service;
 
-import com.example.backend.dto.ProductDto;
-import com.example.backend.dto.ProductEditDto;
-import com.example.backend.entity.Product;
-import com.example.backend.dto.ProductForm;
-import com.example.backend.entity.ProductImage;
-import com.example.backend.repository.ProductImageRepository;
-import com.example.backend.repository.ProductRepository;
+import com.example.backend.product.dto.ProductDto;
+import com.example.backend.product.dto.ProductEditDto;
+import com.example.backend.product.entity.Product;
+import com.example.backend.product.dto.ProductForm;
+import com.example.backend.product.entity.ProductImage;
+import com.example.backend.product.repository.ProductImageRepository;
+import com.example.backend.product.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -46,9 +46,7 @@ public class ProductService {
 
         if (productForm.getImages() != null) {
             for (MultipartFile file : productForm.getImages()) {
-
                 try {
-
                     String s3Url = s3Uploader.upload(file, String.valueOf(product.getId()));
                     ProductImage image = new ProductImage();
                     image.setOriginalFileName(file.getOriginalFilename());
@@ -58,29 +56,9 @@ public class ProductService {
                 } catch (IOException e) {
                     throw new RuntimeException("업로드 실패 : " + e.getMessage(), e);
                 }
-//            String originalFileName = file.getOriginalFilename();
-//            String uuid = UUID.randomUUID().toString();
-//            String storedName = uuid + "_" + originalFileName;
-//            String uploadDir = "C:/Temp/prj4/productFile";
-
-//            File savedFile = new File(uploadDir, storedName);
-//            try {
-//                file.transferTo(savedFile);
-//            } catch (IOException e) {
-//                throw new RuntimeException("파일 저장에 실패했습니다");
-//            }
-//
-//            ProductImage image = new ProductImage();
-//            image.setOriginalFileName(originalFileName);
-//            image.setStoredPath(uploadDir + "/" + storedName);
-//            image.setProduct(product);
-//            imageList.add(image);
-
             }
         }
         productImageRepository.saveAll(imageList);
-
-        // 이미지 DB에 저장
     }
 
     public Map<String, Object> list(Integer pageNumber) {
@@ -139,6 +117,12 @@ public class ProductService {
         productRepository.delete(product);
     }
 
+    // url에서 key만 따오는 메소드
+    private String extractS3Key(String url) {
+        // 예: https://bucket.s3.region.amazonaws.com/123/abc.jpg -> 123/abc.jpg
+        return url.substring(url.indexOf(".com/") + 5);
+    }
+
     public void edit(Long id, ProductEditDto dto) {
         Product product = productRepository.findById(id).get();
         product.setProductName(dto.getProductName());
@@ -146,39 +130,33 @@ public class ProductService {
         product.setCategory(dto.getCategory());
         product.setInfo(dto.getInfo());
         product.setQuantity(dto.getQuantity());
+
+        // 삭제 처리
         if (dto.getDeletedImages() != null) {
             for (String path : dto.getDeletedImages()) {
-                productImageRepository.deleteByStoredPath(path); // or deleteByPath(path)
-                File file = new File("C:/Temp/prj4/productFile", extractFileName(path));
-                file.delete();
+                s3Uploader.delete(extractS3Key(path));
+                productImageRepository.deleteByStoredPath(path);
             }
-            productRepository.save(product);
         }
+        // 파일 저장
         if (dto.getNewImages() != null) {
             List<ProductImage> imageList = new ArrayList<>();
+
             for (MultipartFile file : dto.getNewImages()) {
-                String originalFileName = file.getOriginalFilename();
-                String uuid = UUID.randomUUID().toString();
-                String storedName = uuid + "_" + originalFileName;
-                String uploadDir = "C:/Temp/prj4/productFile";
-
-                File savedFile = new File(uploadDir, storedName);
+                String s3Url = null;
                 try {
-                    file.transferTo(savedFile);
+                    s3Url = s3Uploader.upload(file, String.valueOf(product.getId()));
+                    ProductImage image = new ProductImage();
+                    image.setOriginalFileName(file.getOriginalFilename());
+                    image.setStoredPath(s3Url);
+                    image.setProduct(product);
+                    imageList.add(image);
                 } catch (IOException e) {
-                    throw new RuntimeException("파일 저장 실패");
+                    throw new RuntimeException(e);
                 }
-
-                ProductImage image = new ProductImage();
-                image.setOriginalFileName(originalFileName);
-                image.setStoredPath(uploadDir + "/" + storedName);
-                image.setProduct(product);
-                imageList.add(image);
             }
-
             productImageRepository.saveAll(imageList);
         }
-
     }
 
     private String extractFileName(String path) {
